Backend/requirements.txt
langchain-community
langchain-groq
python-dotenv
fastapi
uvicorn
langgraph